Rename `decorateWithMetrics` to just `metrics`
