+++ lib/Onyx.js
@@ -749,7 +749,7 @@ const Onyx = {
  */
 function applyDecorators() {
     // We're requiring the script dynamically here so that it's only evaluated when decorators are used
-    const decorate = require('./decorateWithMetrics');
+    const decorate = require('./metrics');
 
     // Re-assign with decorated functions
     /* eslint-disable no-func-assign */
