+++ lib/metrics.js
